refactor: extract shared useDragOver hook to eliminate duplication between useFileDrop and useDocumentDrop
The CI duplication checker flagged 36 overlapping lines of drag-event
handling.  Extract the common drag-enter/leave/over/drop + counter logic
into a shared useDragOver hook. Both useFileDrop (ChatInput) and
useDocumentDrop (knowledge base / skill cards) now delegate to it,
keeping only their file-type routing logic.
