+++ ayunis-core-frontend/src/shared/hooks/useDocumentDrop.ts
@@ -1,6 +1,7 @@
-import { useCallback, useEffect, useState, type RefObject } from 'react';
+import { useCallback, type RefObject } from 'react';
 import { useTranslation } from 'react-i18next';
 import { showError } from '@/shared/lib/toast';
+import { useDragOver } from './useDragOver';
 
 interface UseDocumentDropOptions {
   containerRef: RefObject<HTMLElement | null>;
@@ -20,8 +21,6 @@ export function useDocumentDrop({
   disabled = false,
 }: UseDocumentDropOptions): UseDocumentDropResult {
   const { t } = useTranslation('common');
-  const [isDragging, setIsDragging] = useState(false);
-  const [, setDragCounter] = useState(0);
 
   const isValidFile = useCallback(
     (file: File): boolean => {
@@ -31,43 +30,8 @@ export function useDocumentDrop({
     [acceptedExtensions],
   );
 
-  useEffect(() => {
-    const container = containerRef.current;
-    if (!container || disabled) return;
-
-    const handleDragEnter = (e: DragEvent) => {
-      e.preventDefault();
-      e.stopPropagation();
-      setDragCounter((prev) => prev + 1);
-      setIsDragging(true);
-    };
-
-    const handleDragLeave = (e: DragEvent) => {
-      e.preventDefault();
-      e.stopPropagation();
-      setDragCounter((prev) => {
-        const newCount = prev - 1;
-        if (newCount === 0) {
-          setIsDragging(false);
-        }
-        return newCount;
-      });
-    };
-
-    const handleDragOver = (e: DragEvent) => {
-      e.preventDefault();
-      e.stopPropagation();
-    };
-
-    const handleDrop = (e: DragEvent) => {
-      e.preventDefault();
-      e.stopPropagation();
-      setIsDragging(false);
-      setDragCounter(0);
-
-      const files = e.dataTransfer?.files;
-      if (!files || files.length === 0) return;
-
+  const handleDrop = useCallback(
+    (files: FileList) => {
       const validFiles = Array.from(files).filter(isValidFile);
       if (validFiles.length === 0) {
         showError(t('chatInput.invalidDroppedFileType'));
@@ -79,20 +43,15 @@ export function useDocumentDrop({
       }
 
       onDrop(validFiles[0]);
-    };
+    },
+    [isValidFile, onDrop, t],
+  );
 
-    container.addEventListener('dragenter', handleDragEnter);
-    container.addEventListener('dragleave', handleDragLeave);
-    container.addEventListener('dragover', handleDragOver);
-    container.addEventListener('drop', handleDrop);
-
-    return () => {
-      container.removeEventListener('dragenter', handleDragEnter);
-      container.removeEventListener('dragleave', handleDragLeave);
-      container.removeEventListener('dragover', handleDragOver);
-      container.removeEventListener('drop', handleDrop);
-    };
-  }, [containerRef, onDrop, disabled, isValidFile, t]);
+  const { isDragging } = useDragOver({
+    containerRef,
+    onDrop: handleDrop,
+    disabled,
+  });
 
   return { isDragging };
 }
